fix: spa shouldn't use popover script directly
